Move loadFlow endpoints to dedicated controller
